requirements.txt
google-cloud-documentai
google-cloud-storage